requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
spacy==3.6.0
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
albumentations==1.3.1
imutils==0.5.4
scipy==1.11.1
sqlalchemy==2.0.19
python-dateutil==2.8.2
plotly==5.15.0
joblib==1.3.2